+++ CalcCoM_f.c
@@ -8,16 +8,16 @@
 
 
 
-void CalcCoM_f(Struct_uLINK uLINK[],float *com)
+void CalcCoM_f(Struct_uLINK uLINK[],double *com)
 {
-    static float M;
+    static double M;
     static int init_tmp=1;
     if (init_tmp==1)
     {
         M = TotalMass_f(uLINK,1);
         init_tmp=0;
     }
-    //float M;
+    //double M;
     MatClrf(com,3,1);
     //M = TotalMass_f(uLINK,1);
     CalcMC_f(uLINK,com,1);
