+++ CalcCoM_f.h
@@ -2,6 +2,6 @@
 #define H_CalcCoM_f
 
 
-void CalcCoM_f(Struct_uLINK uLINK[],float *com);
+void CalcCoM_f(Struct_uLINK uLINK[],double *com);
 
 #endif
