+++ DrawMarker.c
@@ -12,7 +12,7 @@
 
 void DrawMarker(gsl_vector * com)
 {
-    float d = 0.1;
+    double d = 0.1;
 
     glLineWidth( 1.0f );
     glBegin(GL_LINES);
