+++ Gravity_f.h
@@ -2,7 +2,7 @@
 #define H_Gravity_f
 
 
-void Gravity_f(Struct_uLINK uLINK[],Struct_State *Status, int j, float *f, float *t);
+void Gravity_f(Struct_uLINK uLINK[],Struct_State *Status, int j, double *f, double *t);
 
 
 #endif
