+++ Scenarios.h
@@ -11,14 +11,14 @@
 
 
 /**
- * \fn void Scenario_desired_trajectory(float *qd, double t, int *desired_support, float *distribution_y)
+ * \fn void Scenario_desired_trajectory(double *qd, double t, int *desired_support, double *distribution_y)
  * \brief Compute the articular trajectories for a desired scenario
  *
  * \param uLINK[] Structure wich describe the robot link by link
  * \param com Out: vector containing the CoM position as result
  */
 
-void Scenario_desired_trajectory(float *qd, double t, int *desired_support, float *distribution_y);
+void Scenario_desired_trajectory(double *qd, double t, int *desired_support, double *distribution_y);
