+++ Stabilizator_f.c
@@ -8,9 +8,9 @@
 #include "Setup.h"
 
 
-void Stabilizator_f( Struct_uLINK uLINK[], Struct_State *Status, float *stab)
+void Stabilizator_f( Struct_uLINK uLINK[], Struct_State *Status, double *stab)
 {
-    float com[3];
+    double com[3];
     CalcCoM_f(uLINK,com);
 
 
@@ -36,8 +36,8 @@ void Stabilizator_f( Struct_uLINK uLINK[], Struct_State *Status, float *stab)
         //gsl_vector_set (stab,10, - 100*(gsl_vector_get (Status->FootCenter_R,0) - gsl_vector_get (com,0)) + 20*(gsl_vector_get(com,0)-gsl_vector_get(Status->com_old,0))/Dtime);
         //gsl_vector_set (stab,11, + 200*(gsl_vector_get (Status->FootCenter_R,1) - gsl_vector_get (com,1)) - 20*(gsl_vector_get(com,1)-gsl_vector_get(Status->com_old,1))/Dtime);
 
-        stab[4] = (float)(- 10.f*(Status->FootCenter_R[0] - com[0]) + 2.f*(com[0]-Status->com_old[0])/Te);
-        stab[5] = (float)(  20.f*(Status->FootCenter_R[1] - com[1]) - 2.f*(com[1]-Status->com_old[1])/Te);
+        stab[4] = (double)(- 10.f*(Status->FootCenter_R[0] - com[0]) + 2.f*(com[0]-Status->com_old[0])/Te);
+        stab[5] = (double)(  20.f*(Status->FootCenter_R[1] - com[1]) - 2.f*(com[1]-Status->com_old[1])/Te);
     }
 
     if (Status->desired_support==2)
@@ -49,8 +49,8 @@ void Stabilizator_f( Struct_uLINK uLINK[], Struct_State *Status, float *stab)
         //gsl_vector_set (stab,16, - 100*(gsl_vector_get (Status->FootCenter_L,0) - gsl_vector_get (com,0)) + 20*(gsl_vector_get(com,0)-gsl_vector_get(Status->com_old,0))/Dtime);
         //gsl_vector_set (stab,17, + 200*(gsl_vector_get (Status->FootCenter_L,1) - gsl_vector_get (com,1)) - 20*(gsl_vector_get(com,1)-gsl_vector_get(Status->com_old,1))/Dtime);
 
-        stab[10] = (float)(- 10.f*(Status->FootCenter_L[0] - com[0]) + 2.f*(com[0]-Status->com_old[0])/Te);
-        stab[11] = (float)(  20.f*(Status->FootCenter_L[1] - com[1]) - 2.f*(com[1]-Status->com_old[1])/Te);
+        stab[10] = (double)(- 10.f*(Status->FootCenter_L[0] - com[0]) + 2.f*(com[0]-Status->com_old[0])/Te);
+        stab[11] = (double)(  20.f*(Status->FootCenter_L[1] - com[1]) - 2.f*(com[1]-Status->com_old[1])/Te);
 
 
     }
