+++ StaticTrajectory_f.h
@@ -2,11 +2,11 @@
 #define H_StaticTrajectory_f
 
 
-void StaticTrajectory_f(float *qd, double t, int *desired_support, float *distribution_y);
+void StaticTrajectory_f(double *qd, double t, int *desired_support, double *distribution_y);
 
-void OneFoot_f(float *qd, double t, int *desired_support, float *distribution_y);
+void OneFoot_f(double *qd, double t, int *desired_support, double *distribution_y);
 
-void GravityExperiment_f(float *qd, double t, int *desired_support, float *distribution_y);
+void GravityExperiment_f(double *qd, double t, int *desired_support, double *distribution_y);
 
 
 #endif
