+++ TotalMass_f.h
@@ -2,6 +2,6 @@
 #define H_TotalMass_f
 
 
-float TotalMass_f(Struct_uLINK uLINK[],int j);
+double TotalMass_f(Struct_uLINK uLINK[],int j);
 
 #endif
